+++ sdk/iot/Azure.Iot.Hub.Service/samples/IotHubClientSamples/Program.cs
@@ -28,7 +28,7 @@ namespace Azure.Iot.Hub.Service.Samples
                 });
 
             // Instantiate the client
-            IoTHubServiceClient hubClient = new IoTHubServiceClient(options.IotHubConnectionString);
+            IotHubServiceClient hubClient = new IotHubServiceClient(options.IotHubConnectionString);
 
             // Run the samples
             var deviceIdentityLifecycleSamples = new DeviceIdentityLifecycleSamples(hubClient);