+++ sdk/storage/Azure.Storage.Blobs.ChangeFeed/src/Models/ChangeFeedCursor.cs
@@ -3,7 +3,7 @@
 
 using System;
 
-namespace Azure.Storage.Blobs.ChangeFeed.Models
+namespace Azure.Storage.Blobs.ChangeFeed
 {
     /// <summary>
     /// BlobChangeFeedCursor.