+++ sdk/storage/Azure.Storage.Blobs.ChangeFeed/src/Shard.cs
@@ -5,7 +5,6 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Azure.Storage.Blobs.Models;
-using Azure.Storage.Blobs.ChangeFeed.Models;
 using System.Threading;
 
 namespace Azure.Storage.Blobs.ChangeFeed